requirements.txt
#Libraries

#PyQt6
Pandas
mysql-connector-python
matplotlib
scikit-learn
Seaborn
sqlalchemy
PyQt5
PySide6
QT-PyQt-PySide-Custom-Widgets
pipwin
pipwin install cairocffi
iconify
python-dotenv